requirements.txt
amqp==5.3.1
anthropic==0.69.0
influxdb==5.3.2
metpx-sr3==3.1.0
OWSLib==0.34.1
paho-mqtt==2.1.0
pip==25.2
reactivex==4.0.4
setuptools==80.9.0
wheel==0.45.1
